(chore): move getGitData and uploadGitData to src/helpers/git/instrument-helpers.ts
